Added missing cast.

+++ SimpleTrustZone/CM33_s/main_s.c
@@ -53,7 +53,7 @@ int main(void) {
   /* Get non-secure reset handler */
   NonSecure_ResetHandler = (funcptr_void)(*((uint32_t *)((TZ_START_NS) + 4U)));
  
-  printf("Jump to non-secure application at 0x%08X\n", NonSecure_ResetHandler);
+  printf("Jump to non-secure application at 0x%08X\n", (uint32_t)NonSecure_ResetHandler);
 
   /* Start non-secure state software application */
   NonSecure_ResetHandler();
